Search date examples description
Adding a `orderStatus` content type - it is small, so not a size worry,
and it makes a lot of sense to have the whole cell clickable when only
the sort icon is shown in the cell, which is the case in many of the
search examples. There is a lot of them, so it seems wise to add this
content type.